Trying to read file size
Can we add a check if the file exists first before trying to read the file size? 
We often do DB syncs between environments, without copying all the uploads, just add an .htaccess rewrite, and this poses too many warnings of the filesize failing because not being able to find the file locally.